Safari-611.1.10/LayoutTests/accessibility/mac/unexpected-help-text-expected.txt

This is a checkbox
text
This tests that help text will not be appended by ancestor elements incorrectly. In this example, a table has a summary attribute. The checkbox's help attribute should not be the table summary.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS accessibilityController.focusedElement.helpText is 'AXHelp: '
PASS successfullyParsed is true

TEST COMPLETE
